feat: All classes and Functions have been moved to standalone files in library/
