+++ library/__init__.py
@@ -1,0 +1,11 @@
+from . import argumentparser
+from . import snakecharacter
+from .globalvars import *
+from . import searchalgos
+from . import graphs
+from . import queue
+from . import heuristics
+from . import cfood
+from . import score
+from . import grid
+from . import movement
