+++ library/globalvars.py
@@ -1,3 +1,5 @@
+from typing import Protocol, Dict, List, Iterator, Tuple, TypeVar, Optional
+
 #Global Variables
 
 GRIDSIZE = 20
