+++ library/queue.py
@@ -1,0 +1,29 @@
+import collections
+import heapq
+from .globalvars import *
+
+class Queue:
+    def __init__(self):
+        self.elements = collections.deque()
+    
+    def empty(self) -> bool:
+        return not self.elements
+    
+    def put(self, x: T):
+        self.elements.append(x)
+    
+    def get(self) -> T:
+        return self.elements.popleft()
+    
+class PriorityQueue:
+    def __init__(self):
+        self.elements: List[Tuple[float, T]] = []
+    
+    def empty(self) -> bool:
+        return not self.elements
+    
+    def put(self, item: T, priority: float):
+        heapq.heappush(self.elements, (priority, item))
+    
+    def get(self) -> T:
+        return heapq.heappop(self.elements)[1]
