+++ library/score.py
@@ -1,0 +1,3 @@
+class Score(object):
+    def __init__(self):
+        self.score = 0
